requirements.txt
Flask
Flask-Cors
Flask-Limiter
requests
beautifulsoup4
Flask-JWT-Extended
APScheduler
thefuzz
pywebpush
SQLAlchemy
python-dotenv
pytest
pytest-mock
pytest-flask
cryptography
flasgger
